server_sync/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
aiosqlite==0.20.0
attrs==24.3.0
contourpy==1.3.1
cycler==0.12.1
discord.py==2.4.0
fonttools==4.55.3
frozenlist==1.5.0
greenlet==3.1.1
idna==3.10
kiwisolver==1.4.8
matplotlib==3.10.0
multidict==6.1.0
numpy==2.2.1
packaging==24.2
pillow==11.1.0
propcache==0.2.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
six==1.17.0
SQLAlchemy==2.0.37
typing_extensions==4.12.2
yarl==1.18.3
